requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.1
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
category-encoders==2.6.4